chat/requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.2.1
django-cors-headers==3.0.0
django-rest-framework==0.1.0
django-serializer==0.0.8
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.16.3
pipenv==2018.11.26
psycopg2==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
urllib3==1.25.2
virtualenv==16.5.0
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
wrapt==1.11.1
